+++ src/app/d2-vendors/vendors.tsx
@@ -75,7 +75,7 @@ export default class Vendors extends React.Component<Props, State> {
 
     if (!vendorsResponse || !defs) {
       // TODO: loading component!
-      return <div className="vendor dim-page">Loading...</div>;
+      return <div className="vendor dim-page"><i className="fa fa-spinner fa-spin"/></div>;
     }
 
     return (